Separating concerns for the `KeyHandler` class
The `KeyHandler` class, as it was, made it impossible to use anything
else than files as data store for the keys' data. This patch extracts
the logic common to all key handlers in a separate base class, named
`AbstractKeyHandler`, and makes `KeyHandler` inherit from it.

That way, it is now possible to use other data stores for key handlers,
such as databases; or even to define totally different key handlers,
e.g. one that would compute nonces based on timestamps.

Also added a check for the maximum value of nonces.

Added unit tests on all of that.